agent/pyproject.toml
[project]
name = "vision-quest-agent"
version = "0.1.0"
description = "Vision Quest NPC agents powered by Vision Agents SDK"
requires-python = ">=3.11"
dependencies = [
    "vision-agents[getstream,gemini]",
    "python-dotenv",
    "moondream>=0.2.0",
    "requests>=2.32.5",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
